+++ opentelemetry-sdk/src/propagation/mod.rs
@@ -1,6 +1,8 @@
 //! OpenTelemetry Propagators
 mod baggage;
+mod env;
 mod trace_context;
 
 pub use baggage::BaggagePropagator;
+pub use env::set_global_text_map_propagator_from_env;
 pub use trace_context::TraceContextPropagator;